split objects in dedicated files
